Adds less greedy symbol filtering
Filters (or at least attempts to) only the portion of the text containing the section to format. This should hopefully reduce unexpected behavior for users.

While this will allow users to add in their own server info separators, which is out of scope of this plugin, this change is intended instead to allow for more pretty formatting options. Like adding a map icon symbol in front of the text, a la Orchestrion, for example.